requirements.txt
fastapi
openai
pydub
python-dotenv
pymongo
uvicorn
python-multipart